requirements.txt
pandas
openpyxl
fastapi
uvicorn
requests
dnspython
